Simplified policy_access_rule test
